Refactor from review
