+++ src/actions/images.ts
@@ -10,13 +10,52 @@ import { fileCacheAgentFilter } from "../utils/filters.js";
 import imageGoogleAgent from "../agents/image_google_agent.js";
 import imageOpenaiAgent from "../agents/image_openai_agent.js";
 import { MulmoScriptMethods, Text2ImageAgentInfo } from "../methods/index.js";
-import { imagePreprocessAgent } from "../agents/index.js";
+import { processChart, processMarkdown, processTextSlide, processImage, processMermaid } from "../utils/image_preprocess.js";
 
 const { default: __, ...vanillaAgents } = agents;
 
 dotenv.config();
 // const openai = new OpenAI();
 import { GoogleAuth } from "google-auth-library";
+
+const imagePreprocessAgent = async (namedInputs: {
+  context: MulmoStudioContext;
+  beat: MulmoStudioBeat;
+  index: number;
+  suffix: string;
+  imageDirPath: string;
+  imageAgentInfo: Text2ImageAgentInfo;
+}) => {
+  const { context, beat, index, suffix, imageDirPath, imageAgentInfo } = namedInputs;
+  const imageParams = { ...imageAgentInfo.imageParams, ...beat.imageParams };
+  const prompt = (beat.imagePrompt || beat.text) + "\n" + (imageParams.style || "");
+  const imagePath = `${imageDirPath}/${context.studio.filename}/${index}${suffix}.png`;
+  const aspectRatio = MulmoScriptMethods.getAspectRatio(context.studio.script);
+  const textSlideStyle = MulmoScriptMethods.getTextSlideStyle(context.studio.script, beat);
+
+  if (beat.image) {
+    const canvasSize = MulmoScriptMethods.getCanvasSize(context.studio.script);
+    const processorParams = { beat, context, imagePath, textSlideStyle, canvasSize };
+
+    if (beat.image.type === "textSlide") {
+      await processTextSlide(processorParams);
+    } else if (beat.image.type === "markdown") {
+      await processMarkdown(processorParams);
+    } else if (beat.image.type === "image") {
+      const path = processImage(processorParams);
+      if (path) {
+        // undefined prompt indicates that image generation is not needed
+        return { path, prompt: undefined, imageParams, aspectRatio };
+      }
+    } else if (beat.image.type === "chart") {
+      await processChart(processorParams);
+    } else if (beat.image.type === "mermaid") {
+      await processMermaid(processorParams);
+    }
+  }
+
+  return { path: imagePath, prompt, imageParams, aspectRatio };
+};
 
 const graph_data: GraphData = {
   version: 0.5,
