+++ src/agents/index.ts
@@ -6,7 +6,6 @@ import mulmoPromptsAgent from "./mulmo_prompts_agent.js";
 import ttsNijivoiceAgent from "./tts_nijivoice_agent.js";
 import ttsOpenaiAgent from "./tts_openai_agent.js";
 import validateMulmoScriptAgent from "./validate_mulmo_script_agent.js";
-import { imagePreprocessAgent } from "./image_preprocess_agent.js";
 
 import { browserlessAgent } from "@graphai/browserless_agent";
 import { textInputAgent } from "@graphai/input_agents";
@@ -28,5 +27,4 @@ export {
   ttsNijivoiceAgent,
   ttsOpenaiAgent,
   validateMulmoScriptAgent,
-  imagePreprocessAgent,
 };
